+++ requirements.txt
@@ -1,0 +1,1 @@
+inquirer==2.7.0
